consumer.py: add the consumer code
The trick in this code is around l.58 to l.67 and the related `recv_rp_dat`
function. The later make a recv request with timeout set to the return time,
this timeout is required by the callee logic ; it the simply apply RP_dat to the
data or return None on fail. The do_loop try/except is put to catch the timeout
and ignore the frame.

Signed-off-by: Victor333Huesca <Victor333Huesca@hotmail.com>